Test and read me added to chrome ux report

+++ plugins/chrome-ux-report-backend/migrations/20210315083034_init.js
@@ -20,68 +20,68 @@
  * @param {import('knex').Knex} knex
  */
 exports.up = async function up(knex) {
-    await knex.schema.createTable('origins', table => {
-        table.comment('The table of chrome ux report sites');
-        table.increments('id').primary().notNullable();
-        table.text('origin').unique().notNullable();
-    });
+  await knex.schema.createTable('origins', table => {
+    table.comment('The table of chrome ux report sites');
+    table.increments('id').primary().notNullable();
+    table.text('origin').unique().notNullable();
+  });
 
-    await knex.schema.createTable('periods', table => {
-        table.comment('The table of chrome ux report monthsWithYear');
-        table.increments('id').primary().notNullable();
-        table.text('period').unique().notNullable();
-    });
+  await knex.schema.createTable('periods', table => {
+    table.comment('The table of chrome ux report monthsWithYear');
+    table.increments('id').primary().notNullable();
+    table.text('period').unique().notNullable();
+  });
 
-    await knex.schema.createTable('uxMetrics', table => {
-        table.comment('The table of chrome ux report metrics');
-        table.increments('id').primary().notNullable();
-        table
-            .integer('origin_id')
-            .references('id')
-            .inTable('origins')
-            .notNullable()
-            .onDelete('CASCADE');
-        table
-            .integer('period_id')
-            .references('id')
-            .inTable('periods')
-            .notNullable()
-            .onDelete('CASCADE');
-        table.text('form_factor').notNullable();
-        table.text('connection_type').notNullable();
-        table.decimal('fast_fp').notNullable();
-        table.decimal('avg_fp').notNullable();
-        table.decimal('slow_fp').notNullable();
-        table.decimal('fast_fcp').notNullable();
-        table.decimal('avg_fcp').notNullable();
-        table.decimal('slow_fcp').notNullable();
-        table.decimal('fast_dcl').notNullable();
-        table.decimal('avg_dcl').notNullable();
-        table.decimal('slow_dcl').notNullable();
-        table.decimal('fast_ol').notNullable();
-        table.decimal('avg_ol').notNullable();
-        table.decimal('slow_ol').notNullable();
-        table.decimal('fast_fid').notNullable();
-        table.decimal('avg_fid').notNullable();
-        table.decimal('slow_fid').notNullable();
-        table.decimal('fast_ttfb').notNullable();
-        table.decimal('avg_ttfb').notNullable();
-        table.decimal('slow_ttfb').notNullable();
-        table.decimal('small_cls').notNullable();
-        table.decimal('medium_cls').notNullable();
-        table.decimal('large_cls').notNullable();
-        table.decimal('fast_lcp').notNullable();
-        table.decimal('avg_lcp').notNullable();
-        table.decimal('slow_lcp').notNullable();
-        table.unique(['origin_id', 'period_id'], 'origins_periods_idx');
-    });
+  await knex.schema.createTable('uxMetrics', table => {
+    table.comment('The table of chrome ux report metrics');
+    table.increments('id').primary().notNullable();
+    table
+      .integer('origin_id')
+      .references('id')
+      .inTable('origins')
+      .notNullable()
+      .onDelete('CASCADE');
+    table
+      .integer('period_id')
+      .references('id')
+      .inTable('periods')
+      .notNullable()
+      .onDelete('CASCADE');
+    table.text('form_factor').notNullable();
+    table.text('connection_type').notNullable();
+    table.float('fast_fp').notNullable();
+    table.float('avg_fp').notNullable();
+    table.float('slow_fp').notNullable();
+    table.float('fast_fcp').notNullable();
+    table.float('avg_fcp').notNullable();
+    table.float('slow_fcp').notNullable();
+    table.float('fast_dcl').notNullable();
+    table.float('avg_dcl').notNullable();
+    table.float('slow_dcl').notNullable();
+    table.float('fast_ol').notNullable();
+    table.float('avg_ol').notNullable();
+    table.float('slow_ol').notNullable();
+    table.float('fast_fid').notNullable();
+    table.float('avg_fid').notNullable();
+    table.float('slow_fid').notNullable();
+    table.float('fast_ttfb').notNullable();
+    table.float('avg_ttfb').notNullable();
+    table.float('slow_ttfb').notNullable();
+    table.float('small_cls').notNullable();
+    table.float('medium_cls').notNullable();
+    table.float('large_cls').notNullable();
+    table.float('fast_lcp').notNullable();
+    table.float('avg_lcp').notNullable();
+    table.float('slow_lcp').notNullable();
+    table.unique(['origin_id', 'period_id'], 'origins_periods_idx');
+  });
 };
 
 /**
  * @param {import('knex').Knex} knex
  */
 exports.down = async function down(knex) {
-    await knex.schema.dropTable('uxMetrics');
-    await knex.schema.dropTable('origins');
-    await knex.schema.dropTable('periods');
+  await knex.schema.dropTable('uxMetrics');
+  await knex.schema.dropTable('origins');
+  await knex.schema.dropTable('periods');
 };
